quality.tidy: clear six small bugprone checks (13 findings)
Eleven real fixes, two suppressions that name the contract they protect.

The one that mattered most was not a bug after all: `image_content.cpp` handed
`string_view::data()` to `std::from_chars`, which is range-based and *was* being
given the end pointer -- just as a separately-declared local, which is enough to
defeat the check's pattern match. The `string_view` was ceremony over a
`getenv` result that is NUL-terminated by construction, so it is gone and the
call takes the pointer and a computed end directly.

`document.cpp`'s `if (d_settler_installs > 0 && --d_settler_installs == 0)` is
well-defined -- `&&` sequences the guard ahead of the decrement -- but it is the
exact shape the check exists to discourage, and it is library code, so the
mutation became its own statement rather than a suppression. Same treatment for
the `++n` buried in an IIFE inside a Catch2 `CHECK`.

The two kept duplications:

- `tile_planning.cpp`'s clone pair is `Transient` and `Stale` (not `Fresh` and
  `Transient`, which is what it looks like from the diagnostic's anchor): two
  different rungs of doc 02's degradation ladder -- inexact-but-current-revision
  against prior-revision -- separately selected by `plan_layer`, separately
  asserted, each carrying its own rationale. Merging them would have deleted one
  of the two explanations to save four identical lines.
- `refine_idempotence_golden.t.cpp` compares float pixels bitwise because
  byte-exactness IS the assertion (its `enforces:` trailer is
  16-sdlc-and-quality#byte-exact-goldens). Value equality would quietly stop
  distinguishing -0.0 and NaN payloads.

The five misplaced widening casts were all `static_cast<size_t>(i + 1)` on an
index; each now widens before the arithmetic instead of after it.

Verified against the CI toolchain: across the thirteen changed TUs the six
checks go 13 -> 0 with every other check's count unmoved (78 findings -> 65),
and the full 1219-test suite passes.

+++ src/pool/t/workspace_file.t.cpp
@@ -304,7 +304,7 @@ TEST_CASE("disk-full growth surfaces an error, never an abort") {
   arbc::WorkspaceFileChunkSource& ws = **source;
 
   constexpr std::size_t chunk = 64 * 1024;
-  const off_t header_bytes = file_size(path.str());
+  const off_t header_bytes = static_cast<off_t>(file_size(path.str()));
 
   struct rlimit small = old_limit;
   small.rlim_cur = static_cast<rlim_t>(header_bytes) + chunk + 4096; // room for one chunk
